distribute models in schemas.py
